+++ src/Division/DivisionService.php
@@ -7,6 +7,7 @@
 namespace Wizaplace\SDK\Division;
 
 use GuzzleHttp\Exception\ClientException;
+use GuzzleHttp\RequestOptions;
 use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Exception\FeatureNotEnabled;
 use Wizaplace\SDK\Exception\NotFound;
@@ -34,6 +35,37 @@ class DivisionService extends AbstractService
             $divisionUtils = new DivisionUtils();
 
             return $divisionUtils->getDivisions($this->client->get($url));
+        } catch (ClientException $e) {
+            if ($e->getResponse()->getStatusCode() === 404) {
+                throw new NotFound("The division doesn't exist", $e);
+            }
+            if ($e->getResponse()->getStatusCode() === 501) {
+                throw new FeatureNotEnabled("The feature is not enabled", $e);
+            }
+            throw $e;
+        }
+    }
+
+    /**
+     * Change state of a division
+     *
+     * @param string $code
+     * @param bool   $isEnabled
+     *
+     * @return array
+     * @throws FeatureNotEnabled
+     * @throws NotFound
+     */
+    public function set(string $code, bool $isEnabled)
+    {
+        try {
+            $divisionUtils = new DivisionUtils();
+
+            return $divisionUtils->getDivisions($this->client->patch("divisions/{$code}", [
+                RequestOptions::FORM_PARAMS => [
+                    'is_enabled' => $isEnabled,
+                ],
+            ]));
         } catch (ClientException $e) {
             if ($e->getResponse()->getStatusCode() === 404) {
                 throw new NotFound("The division doesn't exist", $e);